index.html: add app-input and app-example-input meta tags
Add special meta tags to give the webserver an opportunity to
pre-define the input log for the react app.

If app-input meta tag has a link attribute set, the app will attempt
to load the verifier log from there.

The app-example-input meta tag provides a way to modify a link for
"load example log" button.