Fix infinite fetch loop
When loading content from the url param,
we were entering an infinite loop because
the `useEffect` to fetch the content was
dependent upon a function that was getting
re-created when the stored logs were
getting updated. So every time we fetched,
we updated the stored logs, and round
and round we went.