Fix annoying flicker
I saw this with a few logs where the "hint" section
at the bottom (that contains the selected and
hovered line) grows in height which causes the
main section with the log to resize, which can
change what line you're hovering on and then causes
the "hint" section to shrink, which causes
the main section to resize again - and so forth.

This change removes the "Selected Line" at the bottom
and also updates the style so the, now reduced,
hint section can overflow on top of the main content
for really long lines.